requirements.txt
streamlit==1.34.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
nltk>=3.8.1
matplotlib>=3.7.0
seaborn>=0.12.2
reportlab>=4.0.0
pandas>=2.0.0
requests>=2.31.0
markdown2>=2.4.0
xhtml2pdf>=0.2.8
streamlit-lottie>=0.0.4
fpdf2>=2.7.4